Add a prefix to hyperlinks if building for GitHub Pages
We shall be deploying this to GitHub Pages, where the URL is of
the scheme <username>/github.io/<reponame>/. This messes with
hyperlinks starting with / (which is what we are using). So, we
have written a `link-prefix` function that will check whether
Pollen has been started with the environment variable POLLEN set
to the value `GH_PAGES`, and if yes, prefix the link with '/ghalib'
(the repo name).